TpetraKernels: Promote ArithTraits into Kokkos namespace.
The ArithTraits class still lives in the Kokkos::Details namespace.
This commit also promotes it (via a 'using' declaration) into the Kokkos
namespace.  This goes along with the public release of the Kokkos
refactor version of Tpetra, because this class is meant for users as
well as for developers.  (It's the Kokkos analog of
Teuchos::ScalarTraits.)