TpetraKernels: Moved operator<<(ostream&,__float128) out of header.
We have to put operator<<(std::ostream&, __float128) somewhere.  Both
Teuchos and TpetraKernels need it.  I found it easier to put it in
Teuchos, so I took its definition out of Kokkos_ArithTraits.hpp (I put
it into Teuchos_ScalarTraits.hpp in an earlier commit).  However, the
__float128 test still needs it, and doesn't include
Teuchos_ScalarTraits.hpp, so I added a copy of it to the __float128
test.