TpetraKernels: Fixed output array layout for dot & nrm2_squared.
The full specializations of KokkosBlas::{dot, nrm2_squared} used
Kokkos::LayoutLeft as the array layout of the output array.  This is
_not_ the default layout for non-Cuda execution spaces.  As a result,
Tpetra::MultiVector was not picking up the specializations; it was
using the (slower) generic versions.  This commit changes the full
specializations of dot() and nrm2_squared() to use the default array
layout for the output array.  Now Tpetra::MultiVector picks up the
full specializations.