TpetraKernels: Fixed performance issue in KokkosBlas::dot for MV.
Christian and I found out that even for LayoutLeft, unrolling across
rows still helps.  We changed the Scalar=double dot product kernels
accordingly.

We still need to figure out the optimal max unroll length, and
strip-mine if the number of columns is greater than that.  Currently,
the max unroll length is 16.  That is optimal for LayoutRight, for one
platform.  Furthermore, currently we don't strip-mine if the number of
columns is greater than that; instead, we fall back to a non-unrolled
kernel.