Tpetra::MultiVector: Fixed bugs in scale() and update().
My previous commits -- esp. the one that made MultiVector's
constructor actually pay attention to the zeroOut = false case --
exposed a bug in two-argument scale(alpha,A).  This was calling
MultiVector::update(alpha, A, 0), which in turn exposed a bug in
KokkosBlas::axpby when alpha != 0,-1,1 and beta == 0.  This commit
fixes the bug in KokkosBlas::axpby().  It also changes
Tpetra::MultiVector::scale(alpha,A) so that it uses KokkosBlas::scal()
instead of just calling MultiVector::update().