Tpetra: switched to use new spmv interface, fixed related issues
This switches Tpetra to use the new spmv interface. One change to
KokkosSparse CrsMatrix is that its getRow[Const] functions are now
templated on the size_type for the SparseRowView type. This
change was necessary to prevent an up to 30% performance drop
if the Matrix uses size_t as its size_type but could get away
with int because the runtime dimensions don't exceed 2B.