TpetraKernels: Fixed KokkosBlas::axpby to use auto & decltype.
When taking a 1-D View of a column of a 2-D View,
KokkosBlas::Impl::Axpby now uses auto and decltype to let
Kokkos::subview deduce the correct type of the column, rather than
trying to specify the type of the 1-D View explicitly.  That explicit
specification was incorrect for LayoutRight Views.  Using auto and
decltype makes the type of the 1-D Views correct, no matter what the
layout of the 2-D Views.