Kokkos Kernels: Stage HTS (Hybrid Triangular Solver) for use.
HTS is a multithreaded sparse triangular solver. The high-level algorithm
combines two different analysis and solution methods into one: level scheduling,
which can often extract a lot of parallelism from very sparse matrices, and
recursive blocking, which extracts a lot of parallelism from dense matrices. In
the symbolic analysis phase, HTS finds a level schedule for the whole matrix,
then uses a contiguous subset of levels that are useful and discards the
remaining levels. This procedure yields a reordered triangle having a densified
remaining block, to which recursive blocking is applied.

This commit stages HTS for use in Kokkos Kernels. It is in the Experimental
namespace. Both the implementation and the interface may change considerably
over time.

  hts.hpp is the API.
  hts_impl.hpp is the implementation header.
  hts.cpp is the implementation file.
  hts_test.cpp is a unit test.
  Makefile.hts is a reference Makefile for hts_test.