TpetraKernels: work on instantiation logic for SPMV, UPDATE, AXPY
The logic was wrong for SPMV, so that Tpetra::CrsMatrix would end up
recompiling the spmv kernels. On top of it I added the instantiations
for int and long to the kernels package. At some point we might want
to use some ETI types logic instead of just adding them manually.

Also I added a HAVE_TPETRAKERNELS_ETI_ONLY cmake variable, which if
set will disable the definition of the generic variant of SPMV, Update
and AXPY so that only the kernels package explicit instatiated types will
work.