KokkosKernels: Only build sparse mat-vec for Tpetra's exec spaces
Only build sparse matrix-vector multiply ("mat-vec") kernels for
Kokkos execution spaces that Tpetra uses.  Before, we were building
them for all execution spaces that Kokkos enables, whether or not they
are enabled in Tpetra.  This should cut down build times and library
sizes by half, for the common case where both Kokkos::OpenMP and
Kokkos::Serial are enabled, but only one of these (OpenMP by default)
is enabled in Tpetra.

This relates to #700.  For the MPI_RELEASE build discussed in that
issue, tpetra build directory is now down to 155 M.

Also, make sure that the mat-vec declarations match the definitions,
in terms of the set of enabled Scalar types.