Fix #22 (Add KokkosSparse::OrdinalTraits)
KokkosSparse::Impl::getDiagCopyWithOffsets used to depend on
Tpetra::Details::OrdinalTraits, which lives in TpetraCore.  This
dependency is incorrect, because TpetraCore depends on KokkosKernels,
not the other way around.

I fixed this by copying Tpetra::Details::OrdinalTraits into
KokkosKernels as the traits class KokkosSparse::OrdinalTraits, and
changing getDiagCopyWithOffsets to depend on the new traits class.

This commit fixes the following issue:
https://github.com/kokkos/kokkos-kernels/issues/22