KokkosKernels: Update example for machine config
This is a minor update which adds an additional example
line to clarify how machine configuration would be
set. Eventually it's anticipated we'll do something like
this for different modes such as Cuda, threads, etc.

Also hooked up the actual comm size to make the demos
a bit more real.