KokkosKernels: Update so free of Teuchos
Removed the blas1 demo as realized it was probably not a good
basis for the first demos since it's the only one with Teuchos.

Removed the Teuchos links in the other demos to emphasize the
system is Teuchos free.

Next step will be to pick another test and apply the
archiver to that, get it working with different Kokkos modes, etc.