Kokkos-Kernels Apply archiver to test (#38)
This update applies the performance archiver
to the KokkosSparse_spmv.cpp test.

In the current form the archiver closely follows
the original implementation and archives to a yaml
file the values which are printed.

This is in progress and intended as a preliminary
demo of how to apply the archiver to a real test.

Turning on the archiver currently requires:
  Building yaml-cpp
  Setting KOKKOSKERNELS_ENABLE_TPLS="yaml"
  Setting YAML_LIBS AND YAML_PATH

Then KokkosSparse_spmv.exe will archive to a file
called KokkosSparse_spmv.yaml in the same folder
and will enter tests results for different machine
configurations based on the test setting.

On the first run the yaml archive is created. Subsequent
runs will compare values to the original entry.

Note that right now the variance in times is high
and will exceed the tolerance triggering the archive
to indicate a non-passing event.