This commit completes the multi-level memory system ETI and compilation.
On KNLs, to enable explicit HBM and DDR management: run generate makefile with:
--with-memkind=/home/projects/x86-64-knl/memkind/20160811/gcc/4.8.3/

This will add HBWSpace to kokkos devices, and ETI'ed functions will instantiate
for both HostSpace and HBWSpace. I updated these for all blas and sparse functions,
because HBWSpace becomes the default memory space when memkind is enabled, and
instantiating HBWSpace becomes necessity.

For GPUs, if you set KOKKOS_DEVICES=CudaHostPinnedSpace,...
This will generate ETI functions for this memory space as well.
This is only added to sparse functions. This is an optional space,
and "NOT Necessary" to instantiate for blas kernels. I did not add
them for this reason.