D2Color: Refactor
Changing name of execution method in Distance-2 coloring classes
to just be execute since MatrixSquared and the Direct method are
in separate classes anyhow.  It's redundant to have the method
that runs the algorithm be:
- `GraphColorD2_MatrixSquared::color_graph_d2_matrix_squared()`
- `GraphColorD2::color_graph_d2()`
Since we know already what we're executing because of the classtype.

Split the Distance-2 Matrix-Squared version into a separate impl file in
`src/graph/impl`.  The new file is
`src/graph/impl/KokkosGraph_Distance2Color_MatrixSquared_impl.hpp`.

Renamed KokkosGraph files in `src/graph` to be more consistent with the
files in `src/graph/impl` in naming convention and to reduce confusion
due to naming conflict in `perf_test/graph`
- `KokkosGraph_graph_color.hpp` --> `KokkosGraph_GraphColor.hpp`
  - Aligns with `impl/KokkosGraph_GraphColor_impl.hpp`
- `KokkosGraph_graph_color_d2.hpp` --> `KokkosGraph_Distance2Color.hpp`
  - Aligns with `impl/KokkosGraph_Distance2Color*_impl.hpp`

TODO:
- Rename the Direct method class to something that
  is more in-line with whether or not it's a vertex-based
  method, edge based method, etc.