Deterministic coloring: adding bit array variant
I am adding a new version of the algorithm called VBDBIT which implements
a 64 bits, bit array to store the banned colors.
I also found a bug related to decrementing the dependency list, it is now
done atomically and that atomic decrement is more rigorously guarded to
happen only when necessary.