Add COLORING_D2_VB and COLORING_D2_VBTP
Updated the algorithm types to be more in-line with the stuff Mehmet
did for his D1 coloring codes in terms of naming.

- `COLORING_D2_VB` is the same as `COLORING_D2`
  - Will want to deprecate out the `COLORING_D2` flag eventually.
- `COLORING_D2_VBTP` is the vertex based d2 coloring with team-policy
  stuff set up.

The Team-Policy version builds and runs but needs work as it's currently
not performant on OpenMP due to a big increase in the number of phases that are
run.  I haven't tested it much in CUDA