Moved HashmapAccumulator example into example dir
The `example/` dir has not been set up to actually build any examples
so I also added a Makefile in there and updated the generate_makefile
script to add a new build target: `build-example` which will build the
example directory contents.

I also added an option to the generated makefile for `build-all` which
will launch both `build-test` and `build-example` so we can choose if
we want to build just examples or the tests or both.