Deterministic Coloring: reactivating VBD and VBDBIT unit test
I am turning back on these algorithms in the deterministic coloring unit-test as they are now fixed.
After fixing the VBDBIT coloring algorithm I am tunring it back on in the distance 1 coloring unit-test
Small change in VBD/VBDBIT to use the built-in kokkos reducer class.
I think this is more advisable for later maintenance...