Luc: Implementation of a structured spmv algorithm
The implementation hinges on the regularity of the sparsity pattern of the matrix.
It is leveraged to access vector entries without using the view of column indices.
The new kernel is faster for the 5 stencils implemented: 1D, 2D and 3D FD and FE.
A unit test and a performance test are added to assess correctness and performance
of the kernel. Three utility functions are implemented in:

test_common/KokkosKernels_Test_Structured_Matrix.hpp

They return matrices corresponding to the discretization of the Laplace operator on
1D, 2D and 3D grids for FD and FE discretization.