D2 Coloring Cleanup
Commenting out a check I had put in on the Distance-1 graph
coloring handle getter where I threw an error if the user tried
to get the graph-coloring handle before it had been created.
I commented that out for now since the GCHandle use-case in
SPGEMM is to get the pointer before it's created.

I tend to think this pattern opens the door for hard-to-find errors
but baking that into the getter might involve mucking around in
SPGEMM, which is complicated.