Use homebrew's coreutils cp if available on osx
If the `coreutils` package is installed on OSX via Homebrew,
there will be a gnu version of cp installed which does respect
the `-u` option.
This will greatly reduce the time it takes to recompile KokkosKernels
on OSX systems since we won't be updating the timestamp on all the
sources every time we recompile.