Add ostream support to print_1Dview
Create a new version of print_1Dview that takes an ostream parameter
to increase flexibiltiy of where we want the output to go.
i.e., use std::cout or std::cerr for stdout or stderr or send in a
      ofstream object if you want to write to a file.
Also add a separator argument that defaults to " " but can be changed
to "\n" to make a newline as the separator between entries, etc.