Fixed out-of-bounds read in read_mtx()
read_mtx() loads a list of (row,col) pairs, sorts them, and then inserts
into the CRS structure one row at a time. The loop over entries in the
current row only checked if the next entry matches the current row, but NOT
that the end of the pair array has been reached. This caused an invalid
read just after inserting the last entry in the matrix.