Improved Gauss-Seidel: partial col loop unrolling
Imitating SPMV, the non-block Gauss-Seidel TeamPolicy versions now
use a partially unrolled loop over columns. This way, the dot product
sum-reducer has exactly the right width for the current batch of
columns. The batch size has been reduced to 8 (from 16) since
in practice, the #columns is 3 or fewer and there is a template
specialization for the apply logic for every batch size from 1 to 8.