Fix #574 (improve KokkosBlas::dot accuracy for float)
For vectors of `float`, the two-argument overload of KokkosBlas::dot
(that takes two vectors (rank-1 Views) as input, and returns the dot
product result as a scalar value on host) gets inaccurate results on
Mac Clang.  This commit fixes that overload of KokkosBlas::dot so that
if the dot product result type is float resp. complex<float>, the
function uses double resp. complex<double> as the type for
intermediate sums.

This change, when applied to Trilinos in Pull Request
trilinos/Trilinos#6634, fixes Trilinos issue
trilinos/Trilinos#6633.