Changed cmake/kokkoskernels_tribits.cmake to produce KokkosKernels*.cmake files.
Changed CMakeLists.txt so that for an installation with parent, i.e.
add_subdirectory(kokkos-kernels), so that headers were installed in
include/kokkos-kernels consistent with where all of eti files are
installed.

Once installed, the install interface for src/CMakeLists.txt didn't
differentiate between building with a parent or not, so I changed it
to include/kokkos-kernels rather than include for installs with a
parent.

Now, KokkosKernels can be built, installed, built/linked against
through using:

>> add_subdirectory(kokkos-kernels)
>> target_link_libraries(compadre PUBLIC kokkoskernels)