Adding a simpler interface to spadd and spgemm
Attempt at a more user friendly interface for spadd/spgemm.
The new interface only requires three matrices, a handle
and two boolean flag that currently must be set to false.
The new interface also wraps both the symbolic and numeric
calls as users typically want both at least the first time
they run the algorithm.

A couple of bugs related to const vs. non-const in the
numeric phase of both spadd and spgemm are addressed.
The next step in that direction will be to write a new
unit-test that performs the numeric phase multiple times
without calling again the symbolic phase.

Additionally two wiki examples are added to show how this interface
is to be used. Once this PR merges the documentation in the
wiki will be updated to show the new examples with simplified
interface.