Adding a smoke test using some CMake logic
The new test needs Kokkos Kernels to be installed first.
I am not sure how to check that we have performed installation
with CMake but if that can be done then the test could be
guarded so that it does not fail when an install is not
performed.
The test also needs "cmake" to be a valid command (i.e. cmake
is in $PATH), that seems like a reasonable assumption given
that we use CMake to configure the package...