unit-test: manually generate .hpp files per backend to include all tests
With this approach we reduce the complexity of the CMake logic.
We also remove all the backend .cpp files but one.
Two downsides exit with this approach:
  1. all new tests need to be explicitely added to Test_Component.hpp
  2. since all the header are gathered in a unique source file developers
     need to be more careful about redifining functions and classes
     between tests
On the other hand this should also promote the creation of utility headers
to gather implementation of common features useful for multiple tests
instead of copy pasting code which makes maintenance harder.