Fix #1111: spmv tpl instantiations
Use KOKKOS_COMPILE_LIBRARY (true while building library, and false when external app is using prebuilt library)
as the "eti avail" template parameter for SpMV TPL definitions (cuSPARSE and MKL). Before was hardcoded to true,
but this caused undefined reference errors when the TPL was available but ETI was not.