CrsMatrix: deprecate constructor without ncols input
We have only one constructor that does not accept
ncols as an input (except for default constructor)
Additionally that constructor does extra work to
compute that value without really telling the user.
The changes deprecate that behavior and move some
code around to group more logically attributes and
methods in the class.