SpMV: adding support for rocSPARSE TPL
rocSPARSE is declared as a TPL for HIP backend
and hooks are in place to call the TPL when the
library supports it (mode and types restrictions
apply). The unit-tests are passing fine, the
performance test is now too old and outdated
to be updated easily. Further work will be needed
to get that done.