MDF: initial implementation, slow but correct
First implementation of the Minimum Discarded Fill (MDF)
algorithm. It allocates a lot of extra memory and computes
transpose of matrices multiple times... so it is really
bad in terms of performance but it does give a correct
answer to example 2.5 in D'Azevedo's original paper titled
"Ordering methods for preconditioned conjugate gradient
methods applied to unstructured grid problems." in SIAM
Journal on Matrix Analysis and Applications 1992.