SpMV: fixing issues with unit-tests tolerance
Implementing a new formula to determine the level
of accuracy to be expected when checking the correctness
of the SpMV algorithm.

Fudging a bit epsilon to allow tests to pass...

The tensor core examples required a slightly different formula
to take into account blockSize in the calculation of the largest
possible error that can occur during SpMV.
Also needed to add a 2x fudge factor, not very satisfying but
definitely acceptable!