Newton sover: serial on device implementation of Newton's method
This is currently experimental and does not have a polished
public interface so it is kept in experimental.
Will also add a test that solves a linear system which
should converge in a single iteration.

The residual norm and the alpha line search parameters do not need
to be stored as views. Simple stack variables are appropriate and
if needed these local values can be stored in the handle for later
retrival as is the case of the "lastResidual".

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>