MDF: improving performance and adding performance test
The performance test allows to generate random matrices, random
diagonal matrices and to read matrices from file. It collects time
for the handle creation, symbolic phase and numeric phase of the
MDF algorithm.

A small change in the methods names is made to make MDF more uniform
with the rest of the library. The unit-test is improved by checking
the results in L and U against analytical solution.

mostly changing the way the discarded fill is computed at each
factorization step, only selecting rows that were impacted by the
last factorized row.