ROT: adding the ROT function to level 1 BLAS
This shall include all interfaces and implementations for the host
call version of the function. Device calls may be added in later
PRs. The implementation will include TPL support for current GPU
and CPU libraries.
while adding support for cublas I realized that the cos and sin values
should not be complex value even for Crot and Zrot... this will require
rewriting the whole implementation from ETI level down : (
Added cuBLAS support and discovered during testing that cuBLAS
does not work if a device pointer is given for c and s inputs.
Accordingly the TPL layer implementation now copies back and forth
to the host to avoid a segfault.