BLAS: mixed gemm
Adding a reproducer for an issue observed by Sierra developers
when gemm is called on matrices with complex and real values in
the same multiplication.

In quite a few place the check performed in unit-tests are not
computing results in the expected type leading to issues with
round-off errors when compared to internal implementation.