KokkosKernels: Use KOKKOSKERNELS_INCLUDE_DIRECTORIES() (TriBITSPub/TriBITS#429)
This replaces the usage of raw include_directories() with the wrapper
KOKKOSKERNELS_INCLUDE_DIRECTORIES().

The deprecated TriBITS macro include_directories() now issues a CMake
Deprecation warning.  The fix is to use tribits_include_directories() in a
TriBITS build instead.

This is the same patch made in Trilinos PR trilinos/Trilinos#11380.