Perf Tests: adding utilities and instantiation wrapper
The goal of this work is to create a common core infrastructure
for the performance test in order to simplify maintenance.
Here two ideas are introduced:
1. the instantiation wrapper
2. the common input parser

both are trying to capture some of the implementation of our
performance test in generic functions that can be called instead
of duplicating logic around instantiation and command line input
parsing.

The new parsing routine checks the parameter name and that the associated
value can be casted properly. It also add some logic to remove
the arguments from argv and argc once they are parsed properly.