Spgemm non-reuse: unification layer and TPLs (#1678)
* Add unification layer, tests for non-reuse SpGEMM

This version has a simpler interface (A, B) -> C, so the user doesn't
have to manage a handle. The default algorithm (SPGEMM_KK) is always
used. The native implementation just calls symbolic and then numeric.

* Add cusparse 11.0+ spgemm noreuse wrapper

* Fix unused local typedef warning/error

* Add cusparse 10.x spgemm noreuse wrapper

* Formatting

* Remove pointless no-reuse spgemm wrapper for cusparse 10, rocsparse

For these versions, the no-reuse wrapper would be identical to the
symbolic wrapper plus the numeric wrapper, so just call those.

* Add MKL non-reuse spgemm wrapper

* Formatting

* Don't try to call 10 spgemm noreuse from cusparse