spgemm: check A/B graphs don't change (debug only)
In the spgemm handle, use hashes to make sure that the graphs of A and B
don't change between the first symbolic call, and all subsequent
symbolic and numeric calls. This is an O(N) test so it's only enabled in
a debug build.

In spgemm test, trigger these checks on purpose to make sure they work.