add MergeMatrixDiagonal abstraction
Adds an abstraction over the diagonals merge-matrix M of two sorted views A and B,
where M[i,] = A[i] > B[j].
A diagonal of such a matrix is the entries running from the bottom-left to the
top-right (from large I / small J to small I / large J).
MergeMatrixDiagonal presents a 1D-View-Like interface into a specific diagonal
of the merge matrix of two views A and B, allowing the caller to retrieve
the value of entries along the diagonal.