Remove calling tribits_exclude_autotools_files()
KokkosKernels is not even calling the wrapper macro
kokkoskernels_exclude_autotools_files() which calls
tribits_exclude_autotools_files().

KokkosKernels has no autotools files anymore and support for
tribits_exclude_autotools_files() was stripped out of TriBITS because it is
not needed anymore.

There will be a slight hack in Trilinos 'develop' to avoid needing to make
this change on Trilinos 'develop'.  But the next time KokkosKernels is
snapshotted into Trilinos, we can remove that (minor) hack.