Remove non-existant dir from CMake include paths
ode/ in the binary dir doesn't exist as a header directory
(because the ODE component has no auto-generated headers for ETI).
The unit tests might be there, but contain no headers that need to be
found.

This fixes errors when building against a Trilinos installation, where
CMake looks for this directory that doesn't exist.