ODE: adding adaptivity test for RK methods
A few variables were not updated in a logical way
leading to very bad and illogical performance of
the adaptive algorithms. This should now be fixed.
Future work can work into reporting the number of
time steps actually taken and also some discussion
around the inner sub-stepping should happen.

A new unit-test is added to test adaptivity although
a criteria to measure the efficacy of the adaptive
algorithm needs to be added.