CMake: error out in certain case
Graph unit tests are unique in that they use default_scalar for the
KokkosKernelsHandle. So if test-eti-only is ON, but neither float nor
double is instatiated, then error out for the graph unit tests.

Users can still build without float or double if they want, but only if
they turn off tests or the graph component.