Kokkos Kernels: removing old code branches ahead of 4.3.00 release
Some old code branches kept for compatibility with Kokkos Core
versions less than 4.2.00 are dropped after release 4.3.00.
The largest changes are the removal of support the Kokkos print
macro in favor of Kokkos::printf() and the removal of half support
from Kokkos Kernels since it is now in Kokkos Core.