Sparse - SpGEMM: labeling spgemm_symbolic in TPL layer a bit more clearly (#2193)
This just improves the readability of the output from the tools as it now
has a symbolic matching the numeric phase. Previously we only had spgemm
in the label which is a bit confusing as it could be the whole spgemm time
i.e. both symbolic and numeric, additionally we had symbolic in the MKL
path but not in cusparse, rocsparse...