batched - dense: Testing and fixing Serial QR
The serial QR algorithms does not have unit-tests and is failing
for non square matrices. See issue #2328.
This first commit fixes the issue with rectangular matrices and
adds a basic test for that use case. Next will work on adding a
test that exercises the interfaces on multiple matrices of different
sizes within a parallel_for. Finally equivalent tests will be added
for the square case as well.
Fixing unused variable error
It looks like the Left NoTranspose ApplyQ is not doing the correct
thing. Will have a look at that next.

Signed-off-by: Luc <lberge@sandia.gov>
Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>