batched - dense: Testing and fixing Serial QR
The serial QR algorithms does not have unit-tests and is failing
for non square matrices. See issue #2328.
This first commit fixes the issue with rectangular matrices and
adds a basic test for that use case. Next will work on adding a
test that exercises the interfaces on multiple matrices of different
sizes within a parallel_for. Finally equivalent tests will be added
for the square case as well.
Fixing unused variable error
It looks like the Left NoTranspose ApplyQ is not doing the correct
thing. Will have a look at that next.

Spliting the tests a bit better, looking at Q and Qt
together to detect issues with consistency between the two.
Also eventually allows to use GEMM to figure out which one
is wrong.

Final fixes and adjusting the tolerance

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>