Sptrsv stream test fixes
SPTRSV_CUSPARSE algorithm is not supported for streams, so it
made no sense to add it to the list of tested algs inside
test_sptrsv_streams. The result was tri_solve_streams being
a no-op for this algorithm. Somehow, this was not caught until
the block algorithm was being used.

Also, some minor cleanup of alg enum handling in the sptrsv handle.
Use a switch statement with a default to catch unhandled enum vals.
print_algorithm should just use the alg string to avoid a duplicated
switch/ifelseif chain. StringToSPTRSVAlgorithm was not handling
several of the enum vals and also returning strings inconsistent
with the strings in return_algorithm_string. Grep revealed no one
using this function, so I removed it.

I added a check in tri_solve_streams to throw an error if an unsupported
alg is used.

Signed-off-by: James Foucar <jgfouca@sandia.gov>