Fixes while documenting (#2466)
* BLAS - scal: removing check on assignable memory spaces

That check is stricter than required as we will values by reference
to perform copies and won't try to reassign pointers.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* BLAS - rot: check at runtime that X and Y have same extent

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* BLAS - rot: improving static assertions

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* BLAS - rotg: check for non-complex types

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* BLAS - ger: check that matrix stores values as non-const

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* BLAS - trmm: check for valid execution space type.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* BLAS: fix missing semi-colon at end of static_assert

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Applying clang-format

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* More clang-format

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Blas - rot: fixing interface of rot

The cosine coefficient is strictly real while the sine coefficient
can be real or complex leading to a bug in the current API. This
commit should fix that for the native and TPL implementation and
the associated unit-test is also fixed accordingly.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* BLAS - ROT: fixing types for Host TPL calls to ROT function

The types for the arguments c and s are actually different and need
to be appropriately propagated through the TPL layers of the library.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

---------

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>