docs: modifying the workflow to add --require-hases for pip (#2527)
This forces pip to check hashes when it installs our dependencies
as part of the workflow. Also adding hashes for source and binary
versions of the dependencies. Finally modifying the link of the
documentation badge in README so it points to the actual
documentation instead of pointing to the last action run.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>