batched - dense: Testing and fixing Serial QR (#2342)
* batched - dense: Testing and fixing Serial QR

The serial QR algorithms does not have unit-tests and is failing
for non square matrices. See issue #2328.
This first commit fixes the issue with rectangular matrices and
adds a basic test for that use case. Next will work on adding a
test that exercises the interfaces on multiple matrices of different
sizes within a parallel_for. Finally equivalent tests will be added
for the square case as well.
Fixing unused variable error
It looks like the Left NoTranspose ApplyQ is not doing the correct
thing. Will have a look at that next.

Spliting the tests a bit better, looking at Q and Qt
together to detect issues with consistency between the two.
Also eventually allows to use GEMM to figure out which one
is wrong.

Final fixes and adjusting the tolerance

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Removing the stride for the workspace

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Applying clang-format

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Adding checks on inputs

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* More clean-ups

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* clang-format

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Batched - QR: adding static assertion on struct templates

Checking that the side, mode and algo are all valid.
Cleaning up the test by using create_mirror_view_and_copy
when appropriate.

Somehow the semantic of create_mirror_view_and_copy is different
from the regular deep_copy so need to make a few changes after
testing on GPUs...

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Batched - QR: adding check for apply Q on the right side.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* clang-format

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* clang-format

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

---------

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>