rocTPLs: modifying cmake logic to prioritize find_package (#2554)
* ROCBLAS: modifying cmake logic to prioritize find_package

We know first try to use find_package and then pull out the
properties from it. Warn users that we will only support that
mode of detection in the future.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Explicitely state which library header needs to be found by CMake

This seems to help a bit to get the proper include directory.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Fix paths for old and new versions of roc TPLs

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Adding deprecation warning to rocSPARSE and rocSOLVER

Warning users that we will switch to FIND_PACKAGE based logic
to detect roc TPLs. They should modify their configuration
script accordingly.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

---------

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>