osx workflow: removing explicit parallelism levels (#2618)
* osx workflow: removing explicit parallelism levels

Switching to implicit parallelism levels for build and
installation as it may change over time based on github
hardware and we don't want to muck with it everytime...

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Explicitly setting the parallelism level

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Untabify

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Bumping cxx standard to 20 and removing deprecated_code_3

The standard should be raised ahead of the upcoming release 5.0.0
to make sure the compiler is supporting it and also to prepare the
code itself if needed. Additionally removing the old
Kokkos_ENABLE_DEPRECATED_CODE_3 configuration option as it is no
longer supported.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

---------

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>
Signed-off-by: alphataubio <alphataubio@gmail.com>