docs: testing literalinclude as proposed by Yuuichi (#2644)
* docs: testing literalinclude as proposed by Yuuichi

This will allow us to save time by not retyping the examples
manually in the docs. More importantly, since the examples are
tested during CI, it will also keep our documentation up to
date when a change is required : )

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* docs: adding literalinclude to blas kernels with examples

Since this technique works pretty well, let us use it uniformly.
Starting with the blas files and next taking care for sparse and
batched examples.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* docs - examples: adding examples for sparse and graph

This inclues the examples directly from the associated source
files that are tested with CI thus keeping the code more easily
up to date. Removing copyright headers to make things more readable.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* docs: fixing typo

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

---------

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>