Perftests benchmark cleanup (#2639)
* Benchmarks: refactoring our benchmark structure

The idea is to fully decouple the benchmarks from the perf tests
and slowly continue to transition our of the performance tests
as they eventually never allowed us to properly monitor performance
regression/progress in the library...

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Benchmark: renaming file and namespace to avoid conflicts

Again the idea is to eventually remove the perf tests so
there is a temporary duplication of the header file that
contains our testing utilities. The old one will be eventually
removed once the perf_tests are migrated to benchmarks.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Benchmarks: modifying github action and applying clang-format

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Benchmark: removing empty directories

Since perf_tests/{lapack,ode} already exclusively used
benchmarks and not perf_tests, they are removed from the
repo and from the associated perf_tests/CMakeLists.txt

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Benchmarks: updating copyright of blas1 benchmarks

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Fixing spmv_bsr_benchmark linking for std::filesystem

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

---------

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>