Batched - SVD: adding new special matrix test for issue #2650
The issue reports that Batched SVD hangs on a 3x6 matrix.
Adding it to our list of special matrices to test for will avoid
potential future regressions.

Limiting the maximum number of iterations by introducing
a for loop instead of a while loop and an input parameter to pass
a maximum number of iterations.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>