Prefer bit manipulation functions from Kokkos Core
Kokkos provides the same function templates as the C++20 standard
library header <bit>.  We propose to defer to these as they are well
specified and well tested.

Currently we use the `*_builtin version` that is defined in the
`Kokkos::Experimental::` namespace but eventually we should be able to
just call the regular unsuffixed function, once the Core library
refactors them to defer to the builtins when not constant evaluated
(something along the lines of `if consteval { /* constexpr friendly
version */ } else { /* builtin fuction */ }`.

One thing to note is that the standard functions only operate on
unsigned integers.  I opted for specifying the corresponding unsigned
integer type as the argument.

Signed-off-by: Damien L-G <dalg24@gmail.com>